Renamed the main file to mockAsync
mockPromise should just get the mock service from mockAsync. Also synchronized requests.
